kidney/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
dotenv==0.9.9
gunicorn==23.0.0
h11==0.15.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
loadenv==0.1.1
packaging==25.0
pillow==11.1.0
pipreqs==0.4.13
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
requests==2.32.3
service-identity==24.2.0
setuptools==78.1.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.2
whitenoise==6.9.0
yarg==0.1.10
zope.interface==7.2
